requirements.txt
fastapi
uvicorn
pandas
langchain_experimental
langchain_groq
python-dotenv
